# perf: optimizing the patterns set matching by exiting early